Minor code reafctoring for router.js Relates #30
